Dived method cast fixed

+++ Secao8/HeaderFilesExercise/HeaderFilesExercise/Math.cpp
@@ -18,7 +18,7 @@ int Multiply(int a, int b)
 
 float Divide(int a, int b)
 {
-	return a / b;
+	return (float)a / (float)b;
 }
